% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validar.R
\name{busca_en}
\alias{busca_en}
\title{Buscar en una base de referencia}
\usage{
busca_en(df, col_sp, vec)
}
\arguments{
\item{df}{dataframe - base de referencia}

\item{col_sp}{nombre de la columna con la informacion de las especies}

\item{vec}{lista de especies que se desean buscar en.}
}
\value{
tbl
}
\description{
Buscar nombres de una lista en una base de datos de referencia
}
